Refactored the file structure
