--- src/App.css
@@ -1,16 +1,0 @@
-.App {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 1em;
-}
-
-.App-wrapper {
-  width: 80%;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-  text-align: justify;
-}
