+++ src/App.js
@@ -1,26 +1,21 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import './App.css';
-import Layout from './components/Layout';
-import Home from './routes/Home';
-import Calculator from './routes/CalculatorPage';
+import './styles/App.css';
+import NavBar from './components/NavBar';
+import Home from './components/Home';
+import Calculator from './components/Calculator';
 import Quote from './components/Quote';
 
 function App() {
   return (
     <React.StrictMode>
       <BrowserRouter>
-        <div className="App">
-          <div className="App-wrapper">
-            <Routes>
-              <Route path="/" element={<Layout />}>
-                <Route index element={<Home />} />
-                <Route path="calculator" element={<Calculator />} />
-                <Route path="quote" element={<Quote />} />
-              </Route>
-            </Routes>
-          </div>
-        </div>
+        <NavBar />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/calculator" element={<Calculator />} />
+          <Route path="/quote" element={<Quote />} />
+        </Routes>
       </BrowserRouter>
     </React.StrictMode>
   );
