--- src/routes/CalculatorPage.jsx
@@ -1,11 +1,0 @@
-import Calculator from '../components/Calculator';
-import '../styles/Calculator.css';
-
-const CalculatorPage = () => (
-  <div className="calculator-page-container">
-    <p>Let&lsquo;s do some Maths!</p>
-    <Calculator />
-  </div>
-);
-
-export default CalculatorPage;
