--- src/routes/QuotePage.jsx
@@ -1,5 +1,0 @@
-import Quote from '../components/Quote';
-
-const QuotePage = () => <Quote />;
-
-export default QuotePage;
