+++ src/styles/App.css
@@ -1,0 +1,5 @@
+body {
+  padding: 0 10%;
+  font-family: 'Roboto', sans-serif;
+  background-color: #f5f5f5;
+}
