+++ src/styles/Home.css
@@ -1,5 +1,5 @@
 .home-container {
-  width: 100%;
+  width: 95%;
   margin-top: 15vh;
   padding: 1em;
   font-family: 'Pacifico', sans-serif;
